src/ai_model_integration.py
```python
import openai
from openai.api_resources import Completion

class AIModelIntegration:
    def __init__(self):
        self.aiCharacter = None
        self.openai.api_key = 'your-api-key'

    def integrateAIModel(self, model='text-davinci-002'):
        self.aiCharacter = openai.Completion.create(engine=model, prompt="Hello, I am your AI character. How can I assist you today?")

    def interactWithAI(self, message):
        response = self.aiCharacter.create(engine=self.aiCharacter['model'], prompt=message)
        return response.choices[0].text.strip()

    def adjustAICharacter(self, investorPreferences):
        # Adjust the AI character based on investor preferences
        # This is a simplified example, in a real-world scenario, this could involve complex NLP processing
        if 'friendly' in investorPreferences:
            self.aiCharacter = openai.Completion.create(engine=self.aiCharacter['model'], prompt="Hello, I am your friendly AI character. How can I assist you today?")
        elif 'professional' in investorPreferences:
            self.aiCharacter = openai.Completion.create(engine=self.aiCharacter['model'], prompt="Hello, I am your professional AI character. How can I assist you today?")
```
