src/tests/test_ai_model_integration.py
```python
import unittest
from unittest.mock import patch
from src import ai_model_integration

class TestAIModelIntegration(unittest.TestCase):

    @patch('src.ai_model_integration.openai.GPT3')
    def test_integrateAIModel(self, mock_gpt3):
        # Mock GPT-3 model
        mock_gpt3.return_value = 'Mocked GPT-3 Model'

        # Call the function to test
        result = ai_model_integration.integrateAIModel()

        # Assert that the function correctly calls the GPT-3 model
        mock_gpt3.assert_called_once()
        self.assertEqual(result, 'Mocked GPT-3 Model')

    @patch('src.ai_model_integration.openai.GPT3')
    def test_interactWithAI(self, mock_gpt3):
        # Mock GPT-3 model
        mock_gpt3.return_value.generate.return_value = 'Hello, I am your AI character.'

        # Call the function to test
        aiCharacter = 'AI Character'
        result = ai_model_integration.interactWithAI(aiCharacter)

        # Assert that the function correctly interacts with the AI character
        mock_gpt3.return_value.generate.assert_called_once_with(aiCharacter)
        self.assertEqual(result, 'Hello, I am your AI character.')

if __name__ == '__main__':
    unittest.main()
```